# docs/openapi/admin.yaml
openapi: 3.0.3
...
info:
  title: voxeltoad — Admin API
  version: 0.1.0
  description: >
    Management-plane REST API for providers, models, routes, plugins, tenants,
    API keys, quotas, usage, and audit. Consumed by the Control Panel UI and the
    generated TypeScript admin SDK.
servers:
  - url: /
    description: Same-origin (admin plane root)
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Operator login → session token
      description: >
        Authenticates email+password and issues an opaque bearer session token
        (TTL 12h). Enforces a per-email failed-attempt lockout (429).
        Unauthenticated endpoint.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Session issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  # ---------------------------------------------------------------------------
  # Global config CRUD (super-admin). Cross-references validated at write time
  # (ADR-0014). Every mutation is audited.
  # ---------------------------------------------------------------------------
  /api/v1/providers:
    get:
      tags: [config]
      summary: List providers
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Provider list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderList'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [config]
      summary: Create or upsert a provider
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Provider' }
      responses:
        '201':
          description: Provider upserted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Provider' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/models:
    get:
      tags: [config]
      summary: List models (accessible by super-admin and tenant-admin)
      description: >
        Returns all model aliases. Accessible by any authenticated operator
        (super-admin or tenant-admin). Models are global shared config;
        writes (POST/PATCH/DELETE) remain super-admin only.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Model list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModelList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [config]
      summary: Create or upsert a model (upstream providers must exist)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Model' }
      responses:
        '201':
          description: Model upserted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Model' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/routes:
    get:
      tags: [config]
      summary: List routes
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Route list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [config]
      summary: Create or upsert a route (candidate providers must exist and each must be an upstream of the referenced model)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Route' }
      responses:
        '201':
          description: Route upserted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Route' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/plugins:
    get:
      tags: [config]
      summary: List plugins
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Plugin list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PluginList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [config]
      summary: Create or upsert a plugin config
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PluginConfig' }
      responses:
        '201':
          description: Plugin upserted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PluginConfig' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/api-keys:
    get:
      tags: [api-keys]
      summary: List the caller tenant's active API keys
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: API key list (non-secret)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIKeyList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [api-keys]
      summary: Create an API key (plaintext returned once)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAPIKeyRequest' }
      responses:
        '201':
          description: Key created; plaintext returned exactly once
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateAPIKeyResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/quotas:
    get:
      tags: [quotas]
      summary: Read a quota balance by scope
      parameters:
        - name: scope
          in: query
          required: true
          schema: { type: string }
          description: >
            Quota scope key, e.g. "tenant:acme", "group:acme/team-a",
            "key:sk-abc". tenant-admin may only query scopes within its tenant.
      responses:
        '200':
          description: Quota balance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuotaBalance' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/groups:
    get:
      tags: [tenancy]
      summary: List the caller tenant's groups
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Group list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GroupList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [tenancy]
      summary: Create a group in the caller's tenant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateGroupRequest' }
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/me:
    get:
      tags: [auth]
      summary: Return the authenticated operator's identity
      description: >
        Returns the operator's id, email, role, and (for tenant-admin)
        tenant_name. Used by the control panel to gate navigation per role
        and to construct quota scope strings.
      responses:
        '200':
          description: Authenticated operator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operator' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ---------------------------------------------------------------------------
  # Roles & Permissions. Phase-2 RBAC custom roles management.
  # super-admin only (wildcard); mounted on the global config group.
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque operator session token from POST /auth/login.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
      description: Opaque keyset cursor from a prior response's next_cursor. Omit for the first page.
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      description: Max items to return.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing/invalid/expired session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not authorized for this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found (or not in caller's tenant)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    TooManyRequests:
      description: Rate limited (e.g. login lockout)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message: { type: string }
            type:
              type: string
              description: OpenAI-compatible error type.
              enum:
                - invalid_request_error
                - authentication_error
                - permission_error
                - not_found
                - rate_limit_error
                - api_error
    ListEnvelope:
      type: object
      description: Uniform list envelope. next_cursor="" means no further pages.
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items: {}
        next_cursor:
          type: string
...
    Pricing:
      type: object
      description: Per-million-token rates in int64 micro-units of Currency (ADR-0013).
      properties:
        prompt_per_1m: { type: integer, format: int64 }
        completion_per_1m: { type: integer, format: int64 }
        currency: { type: string }
        cache_hit_multiplier:
          type: integer
          format: int64
          description: Billing multiplier applied to cached prompt tokens (usage.cached_tokens / Claude cache_read_input_tokens), in int64 micro-units. 1_000_000 = full price, 500_000 = half, 100_000 = 10%. A value of 0 (or field absent) is treated as full price — legacy-safe. Operators wanting a cache discount must set a non-zero value. Per-(alias, provider).
          default: 0
          minimum: 0
          maximum: 1000000

    # --- Auth ---
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    LoginResponse:
      type: object
      required: [token]
      properties:
        token: { type: string, description: Opaque bearer session token (TTL 12h). }

    # --- Config resources (mirror internal/config/schema.go) ---
    ProviderTimeouts:
      type: object
      properties:
        connect: { type: integer, format: int64, description: Nanoseconds (Go time.Duration). }
        first_byte: { type: integer, format: int64 }
        overall: { type: integer, format: int64 }
    Provider:
      type: object
      required: [name]
      properties:
        name: { type: string }
        type: { type: string, description: Descriptive brand (openai/tencent/zhipu/anthropic); does not select adapter. }
        adapter: { type: string, description: Protocol adapter from the registry (openai|claude). }
        base_url: { type: string }
        api_key_ref: { type: string, description: 'Credential ref: env://VAR | db://provider/<name> | plain://literal | bare literal. Masked in responses (env://***). See ADR-0031.' }
        api_key: { type: string, description: 'Write-only plaintext upstream key. When supplied, encrypted and stored in provider_credentials; api_key_ref is set to db://provider/<name>. Never returned.' }
        timeouts: { $ref: '#/components/schemas/ProviderTimeouts' }
        weight: { type: integer }
...
    ModelUpstream:
      type: object
      required: [provider, upstream_model]
      properties:
        provider: { type: string }
        upstream_model: { type: string }
        default_max_tokens: { type: integer }
        pricing: { $ref: '#/components/schemas/Pricing' }
    Model:
      type: object
      required: [alias]
      properties:
        alias: { type: string }
        description: { type: string, description: 'Human-readable summary (model catalog). Informational only.' }
        context_length: { type: integer, description: 'Advertised max context window in tokens. Informational only.' }
        capabilities:
          type: array
          items: { type: string }
          description: 'e.g. [vision, function_calling, streaming]. Surfaced for catalog filtering.'
        tags:
          type: array
          items: { type: string }
          description: Free-form labels for categorization/search.
        upstreams:
          type: array
          items: { $ref: '#/components/schemas/ModelUpstream' }
...
    RouteProvider:
      type: object
      required: [name]
      properties:
        name: { type: string }
        weight: { type: integer }
    Route:
      type: object
      required: [model_alias]
      properties:
        model_alias: { type: string }
        providers:
          type: array
          items: { $ref: '#/components/schemas/RouteProvider' }
        strategy:
          type: string
          enum: [priority, weighted, round_robin, session_affinity]
...
    PluginConfig:
      type: object
      required: [name]
      properties:
        name: { type: string }
        phase: { type: string, enum: [pre, post] }
        params: { type: object, additionalProperties: true }
        enabled: { type: boolean }
        scope: { type: string }
...
    Group:
      type: object
      required: [id, name, enabled]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        enabled: { type: boolean }
    CreateGroupRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
...
    Operator:
      type: object
      description: Non-secret operator view — never includes the password hash.
      required: [id, email, role]
      properties:
        id: { type: integer, format: int64 }
        email: { type: string, format: email }
        role:
          type: string
          enum: [super-admin, tenant-admin]
          deprecated: true
          description: Legacy role text. Prefer role_id for Phase-2 RBAC.
        role_id:
          type: integer
          format: int64
          description: FK to roles table (Phase-2 RBAC).
        scope_kind:
          type: string
          enum: [global, tenant]
          description: Whether the operator's role is global or tenant-scoped.
        permissions:
          type: array
          items: { type: string }
          description: Loaded permission set for this operator's role.
        tenant_id:
          type: integer
          format: int64
          nullable: true
          description: Bound tenant; null for a global-scope operator.
        tenant_name:
          type: string
          nullable: true
          description: >
            Resolved tenant name (for quota scope construction). null for
            super-admin.
...
    APIKeyInfo:
      type: object
      required: [key_id, revoked]
      properties:
        key_id: { type: string }
        revoked: { type: boolean }
        allowed_models:
          type: array
          items: { type: string }
          description: Models this key is restricted to (empty = all).
    CreateAPIKeyRequest:
      type: object
      required: [key_id]
      properties:
        key_id: { type: string }
        allowed_models:
          type: array
          items: { type: string }
          description: Empty/absent = all models allowed.
    CreateAPIKeyResponse:
      type: object
      required: [key_id, api_key]
      properties:
        key_id: { type: string }
        api_key: { type: string, description: "Plaintext key — returned exactly once, never retrievable again." }
...
    QuotaBalance:
      type: object
      required: [scope, balance, currency]
      properties:
        scope: { type: string }
        balance: { type: integer, format: int64, description: Micro-units. }
        currency: { type: string }
...
    ProviderList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/Provider' } }
    ModelList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/Model' } }
    RouteList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/Route' } }
    PluginList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/PluginConfig' } }
...
    GroupList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/Group' } }
...
    APIKeyList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/APIKeyInfo' } }
...